Fix tests for new layout
